packages/sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-test-lab"
version = "0.1.0"
description = "Python SDK and CLI for LLM Test Lab"
requires-python = ">=3.10"
dependencies = [
  "llm-test-lab-core>=0.1.0",
  "typer>=0.9.4",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.0",
  "anyio>=4.0.0",
]

[project.scripts]
llm-test-lab = "llm_test_lab.cli:main"

[tool.setuptools.packages.find]
where = ["."]
